test: move expectGroupHeadingsToBe() to CustomMatchersForGrouping
