+++ tests/CustomMatchers/CustomMatchersForGrouping.ts
@@ -1,4 +1,7 @@
+import { TaskGroups } from '../../src/Query/TaskGroups';
 import type { Field } from '../../src/Query/Filter/Field';
+import type { Task } from '../../src/Task';
+import type { Grouper } from '../../src/Query/Grouper';
 
 declare global {
     namespace jest {
@@ -40,4 +43,20 @@ export function toSupportGroupingWithProperty(field: Field, property: string) {
             `'${field.fieldName()}' field supports grouping, grouping property set to '${fieldGrouper.property}'.`,
         pass: true,
     };
+}
+
+export function expectGroupHeadingsToBe(grouper: Grouper, tasks: Task[], expectedGroupHeadings: string[]) {
+    tasks.sort(() => Math.random() - 0.5);
+
+    // Act
+    const groups = new TaskGroups([grouper], tasks);
+    const groupHeadings: string[] = [];
+    groups.groups.forEach((taskGroup) => {
+        taskGroup.groupHeadings.forEach((heading) => {
+            groupHeadings.push(heading.displayName);
+        });
+    });
+
+    // Assert
+    expect(groupHeadings).toEqual(expectedGroupHeadings);
 }
